serilization fixes

+++ automation/src/main/java/com/vigil/automation/entitity/cucumber/Location.java
@@ -1,5 +1,6 @@
 package com.vigil.automation.entitity.cucumber;
 
+import java.util.Optional;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -7,7 +8,10 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class Location{
-    public int line;
-    public int column;
+public class Location {
+
+   public int line;
+   public Long column;
+
+
 }
